Request Sınıfı Projeye Dahil  Edildi

+++ app/Http/Controllers/Admin/UserController.php
@@ -2,164 +2,128 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-use App\Models\User;
-use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Validator;
-use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
-
+use Illuminate\Support\Facades\Hash;
+use App\Http\Controllers\Controller;
+use App\Http\Requests\Admin\AdminCreateUserRequest;
+use App\Http\Requests\Admin\AdminUpdateUserRequest;
+use App\Http\Requests\Admin\AdminUpdateUserRoleRequest;
+use App\Models\Role;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Illuminate\Support\Str;
+use App\Models\User;
 
 class UserController extends Controller
 {
-    protected $userHiddens = ['email', 'email_verified_at', 'created_at', 'updated_at', 'role_id', 'deleted_at', 'phone_number', 'address', 'city_id', 'google_id', 'github_id', 'pivot', 'role_id'];
 
-    public function users()
+    public function index()
     {
-        $users = User::paginate(10);
+        $users = User::paginate($this->getPerPage());
         $users->makeVisible($this->userHiddens);
-        return response()->json($users, 200);
+        return response()->json($users, JsonResponse::HTTP_OK, [], JSON_UNESCAPED_UNICODE);
     }
 
 
-    public function user($id)
+    public function show($id)
     {
         $user = User::find($id);
         if (!$user) {
-            return response()->json(['error' => 'Kullanıcı bulunamadı.'], 404);
+            return response()->json(['error' => 'Kullanıcı bulunamadı.'], JsonResponse::HTTP_NOT_FOUND, [], JSON_UNESCAPED_UNICODE);
         }
         $user->makeVisible($this->userHiddens);
-        return response()->json($user, 200);
+        return response()->json($user, JsonResponse::HTTP_OK, [], JSON_UNESCAPED_UNICODE);
     }
 
-    public function createUser(Request $request): JsonResponse
+    public function createUser(AdminCreateUserRequest $request): JsonResponse
     {
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|string',
-            'email' => 'nullable|email|unique:users',
-            'phone_number' => 'nullable|unique:users',
-            'address' => 'nullable',
-            'city_id' => 'nullable|integer|exists:cities,id',
-            'description' => 'nullable',
-            'profile_photo' => 'nullable',
-            'password' => 'required|string|min:8|confirmed',
-        ]);
-
-        if ($validator->fails()) {
-            return response()->json(['error' => $validator->errors()], 422);
-        }
-
         $user = new User();
         $user->name = $request->name;
-        $user->email = $request->email ?? null;
+        $user->surname = $request->surname;
+        $user->email = $request->email;
+
         $user->phone_number = $request->phone_number ?? null;
         $user->address = $request->address ?? null;
         $user->city_id = $request->city_id ?? null;
-        $user->description = $request->description ?? null;
-        $user->password = Hash::make($request->password);
+        $user->phone_number = $request->phone_number ?? null;
+        $user->trust_score = $request->trust_score ?? null;
+        $user->role_id = $request->role_id ?? Role::where('slug', 'user')->first()->id;
+        $user->password = bcrypt($request->password);
         $user->save();
 
-        if ($request->hasFile('profile_photo')) {
-            $user->profile_photo = Storage::putFile('public/profile_photos', $request->file('profile_photo'));
+        $id = $user->id;
+        if ($request->hasFile('profile_photo_path')) {
+            $photo = $request->file('profile_photo_path');
+            if (Storage::exists($user->profile_photo_path)) {
+                Storage::delete($user->profile_photo_path);
+            }
+            $photo_name = time() . '.' . $photo->getClientOriginalExtension();
+            $slugName = Str::slug($user->name);
+            $photo->storeAs("public/user-photos/$id-$slugName/", $photo_name);
+
+            $user->profile_photo_path = "user/photos/$id-$slugName/" . $photo_name;
             $user->save();
         }
 
         $user->makeVisible($this->userHiddens);
-
-        return response()->json($user, 201);
+        return response()->json($user, JsonResponse::HTTP_CREATED, [], JSON_UNESCAPED_UNICODE);
     }
 
 
-    public function updateUser(Request $request, $id)
+    public function updateUser(AdminUpdateUserRequest $request, $id)
     {
-        $user = User::findOrFail($id);
-
-        $validator = Validator::make($request->all(), [
-            'name' => ['required', 'string', 'max:255', 'unique:users,name,' . $user->id],
-            'email' => ['required', 'string', 'email', 'max:255', 'unique:users,email,' . $user->id],
-            'phone_number' => ['required', 'string', 'max:255', 'unique:users,phone_number,' . $user->id],
-            'address' => ['required', 'string', 'max:255', 'unique:users,address,' . $user->id],
-            'city_id' => ['required', 'integer', 'exists:cities,id'],
-            'description' => ['required', 'string', 'max:255'],
-            'profile_photo' => ['nullable', 'image', 'mimes:jpeg,png,jpg,gif,svg', 'max:2048'],
-            'password' => ['required', 'min:8', 'confirmed'],
-        ]);
-
-        if ($validator->fails()) {
-            return response()->json(['error' => $validator->errors()], 400);
+        $user = User::find($id);
+        if (!$user) {
+            return response()->json(['error' => 'Kullanıcı bulunamadı.'], JsonResponse::HTTP_NOT_FOUND, [], JSON_UNESCAPED_UNICODE);
         }
 
         $user->name = $request->name;
-
-        if ($request->hasFile('profile_photo')) {
-            $profile_photo = $request->file('profile_photo');
-            Storage::delete($user->profile_photo);
-            $profile_photoName = time() . '.' . $profile_photo->getClientOriginalExtension();
-            $profile_photo->storeAs('public/profile_photos', $profile_photoName);
-            $user->profile_photo = $profile_photoName;
-        }
-
+        $user->surname = $request->surname;
         $user->email = $request->email;
 
-        $user->phone_number = $request->phone_number;
-
-        $user->address = $request->address;
-
-        $user->city_id = $request->city_id;
-
-        $user->description = $request->description;
-
-        if ($request->password != null) {
-            $user->password = Hash::make($request->password);
-        }
-        if ($request->role != null) {
-            $user->role = $request->role;
-        }
-
-        if ($request->club_id != null) {
-            $user->club_id = $request->club_id;
-        }
-
-        if ($request->profile_photo_path != null) {
-            $user->profile_photo_path = $request->profile_photo_path;
-        }
-        if ($request->password != null) {
-            $user->password = Hash::make($request->password);
-        }
-
-        if ($request->hasFile('profile_photo_path')) {
-            $profile_photo_path = $request->file('profile_photo_path');
-            Storage::delete($user->profile_photo_path);
-            $profile_photo_pathName = time() . '.' . $profile_photo_path->getClientOriginalExtension();
-            $profile_photo_path->storeAs('public/profile_photos', $profile_photo_pathName);
-            $user->profile_photo_path = $profile_photo_pathName;
-        }
+        $user->phone_number = $request->phone_number ?? null;
+        $user->address = $request->address ?? null;
+        $user->city_id = $request->city_id ?? null;
+        $user->phone_number = $request->phone_number ?? null;
+        $user->trust_score = $request->trust_score ?? null;
+        $user->role_id = $request->role_id ?? Role::where('slug', 'user')->first()->id;
+        $user->password = bcrypt($request->password);
         $user->save();
+
+        $id = $user->id;
+        if ($request->hasFile('profile_photo_path')) {
+            $photo = $request->file('profile_photo_path');
+            if (Storage::exists($user->profile_photo_path)) {
+                Storage::delete($user->profile_photo_path);
+            }
+            $photo_name = time() . '.' . $photo->getClientOriginalExtension();
+            $slugName = Str::slug($user->name);
+            $photo->storeAs("public/user-photos/$id-$slugName/", $photo_name);
+
+            $user->profile_photo_path = "user/photos/$id-$slugName/" . $photo_name;
+            $user->save();
+        }
+
+        $user->makeVisible($this->userHiddens);
+        return response()->json($user, JsonResponse::HTTP_OK, [], JSON_UNESCAPED_UNICODE);
     }
 
     /**
      * Update user role.
-     * @param Request $request
+     * @param AdminUpdateUserRoleRequest $request
      * @param int $id
      * @return JsonResponse
      */
-    public function updateRole(Request $request, $id): JsonResponse
+    public function updateRole(AdminUpdateUserRoleRequest $request, $id): JsonResponse
     {
-        $user = User::findOrFail($id);
-
-        $validator = Validator::make($request->all(), [
-            'role_id' => ['required', 'exists:roles,id'],
-        ]);
-
-        if ($validator->fails()) {
-            return response()->json(['error' => 'Geçersiz rol'], 400);
+        $user = User::find($id);
+        if (!$user) {
+            return response()->json(['error' => 'Kullanıcı bulunamadı.'], JsonResponse::HTTP_NOT_FOUND, [], JSON_UNESCAPED_UNICODE);
         }
-
         $user->role_id = $request->role_id;
         $user->save();
-
-        return response()->json(['message' => 'Kullanıcının rolü başarıyla güncellendi.'], 200);
+        return response()->json(['message' => 'Kullanıcının rolü başarıyla güncellendi.'], JsonResponse::HTTP_OK, [], JSON_UNESCAPED_UNICODE);
     }
 
 
@@ -170,13 +134,16 @@ class UserController extends Controller
      */
     public function deleteUser($id): JsonResponse
     {
-        $user = User::withTrashed()->findOrFail($id);
+        $user = User::withTrashed()->find($id);
+        if (!$user) {
+            return response()->json(['error' => 'Kullanıcı bulunamadı.'], JsonResponse::HTTP_NOT_FOUND, [], JSON_UNESCAPED_UNICODE);
+        }
 
         if ($user->trashed()) {
-            return response()->json(['error' => 'Kullanıcı zaten silinmiş.'], 400);
+            return response()->json(['error' => 'Kullanıcı zaten silinmiş.'], JsonResponse::HTTP_BAD_REQUEST, [], JSON_UNESCAPED_UNICODE);
         }
         $user->delete();
-        return response()->json(['message' => 'Kullanıcı başarıyla silindi.'], 200);
+        return response()->json(['message' => 'Kullanıcı başarıyla silindi.'], JsonResponse::HTTP_OK, [], JSON_UNESCAPED_UNICODE);
     }
 
     /**
@@ -188,16 +155,17 @@ class UserController extends Controller
     {
         $user = User::withTrashed()->find($id);
         if (!$user) {
-            return response()->json(['error' => 'Kullanıcı bulunamadı'], 404);
+            return response()->json(['error' => 'Kullanıcı bulunamadı'], JsonResponse::HTTP_NOT_FOUND);
         }
 
-        if ($user->deleted_at === null) {
-            return response()->json(['error' => 'Kullanıcı zaten aktif'], 400);
+        if ($user->trashed()) {
+            $user->restore();
+            return response()->json(['message' => 'Kullanıcı başarıyla geri yüklendi'], JsonResponse::HTTP_OK);
+        } else {
+            return response()->json(['error' => 'Kullanıcı zaten aktif'], JsonResponse::HTTP_BAD_REQUEST);
         }
-
-        $user->restore();
-        return response()->json(['message' => 'Kullanıcı başarıyla geri yüklendi'], 200);
     }
+
 
     /**
      * Get all deleted users.
@@ -205,9 +173,8 @@ class UserController extends Controller
      */
     public function deletedUsers(): JsonResponse
     {
-        $users = User::onlyTrashed()->paginate(6);
-        $users->makeVisible($this->userHiddens); // makeVisible() methodu ile gizli alanları görünür hale getirdim. Editör hata verebilir
-        return response()->json($users, 200);
+        $users = User::onlyTrashed()->paginate($this->getPerPage());
+        return response()->json($users, JsonResponse::HTTP_OK, [], JSON_UNESCAPED_UNICODE);
     }
 
     /**
@@ -219,9 +186,9 @@ class UserController extends Controller
     {
         $deletedUser = User::onlyTrashed()->find($id);
         if (!$deletedUser) {
-            return response()->json(['error' => 'Kullanıcı bulunamadı'], 404);
+            return response()->json(['error' => 'Kullanıcı bulunamadı'], JsonResponse::HTTP_NOT_FOUND, [], JSON_UNESCAPED_UNICODE);
         }
         $deletedUser->makeVisible($this->userHiddens);
-        return response()->json($deletedUser, 200);
+        return response()->json($deletedUser, JsonResponse::HTTP_OK, [], JSON_UNESCAPED_UNICODE);
     }
 }
